requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-heroku==0.3.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
kiwisolver==1.1.0
line-bot-sdk==1.15.0
numpy==1.18.1
oauthlib==3.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
